fix(types): long double is its own type again (x87 80-bit, sizeof 16)
`long double` lexed straight to the DOUBLE DataDef. sizeof reported 8 against
16 on x86-64, printf("%Lg") printed nan because glibc reads 80 bits off the
varargs stack, and the mangler emitted Itanium `e` for a value passed as a `d`
— an ABI mismatch on any mangled-direct call taking a long double, which the
C++ library work makes newly reachable.

CAUSE, per the owner and corroborated by the shape of the code: the original
asmjit x86-64 JIT could not emit the x87 80-bit type, so the front end
substituted a double to let such source compile at all. Every other layer was
built for the real thing — DataType::dtLDOUBLE exists, is_real() accounts for
it, madc_mangle emits `e`, the generated macro table advertises __LDBL_MAX__
1.189e+4932, and the builtin registry declares copysignl with long-double
parameters. That is the signature of a narrow backend workaround, not a design
decision. asmjit was then removed, and CIR/c2mir/MIR supports the type fully
(MIR_T_LD, N_LD, new_ld_node, strtold) — but nothing marked the mapping as
provisional, so it survived the backend swap and quietly became wrong. Worth
generalizing: when a backend constraint is lifted, grep for the decisions that
cite it. An audit of the remaining asmjit references turned up two live
behaviours still justified by asmjit (TokenFunc::is_overridden dedup, lambda
hoisting); both are still correct for other reasons and only their comments are
stale. This was the one that had gone wrong.

Two halves, and the second matters as much as the first.

TYPE: a real DataDefLDOUBLE/ddLDOUBLE at size 16, one mapping site in the
type-specifier accumulator, and CIR emitting the specifier PAIR
{N_LONG, N_DOUBLE} — c2mir has no single long-double specifier node because C
has no single token for it, the same way dtUINT64 emits {N_UNSIGNED, N_LONG}.
Three fan-out sites gained companions so nothing regresses now that dtLDOUBLE
actually reaches them: the DataType->DataDef map, and the throw/catch pair,
where a long double rides the double path exactly as it did when it WAS one
(the exception runtime carries a single real payload width).

VALUE: TokenReal stored a double and literals were parsed with strtod, so
`1.0L` was truncated at lex time — before any type could matter, and beyond any
later stage's power to recover. It now stores a long double behind ldval(),
literals parse with strtold, the `L` suffix selects ddLDOUBLE the way `f`
selects ddFLOAT, and CIR emits c2mir's N_LD constant node instead of narrowing
through dval(). Without this half the type fix alone still printed nan for a
long-double LITERAL as a vararg (emitted as an 8-byte node, read as 16) and
folded 1.0L/3.0L at double precision, so the same expression gave one answer
folded and another through a cast that forced runtime arithmetic.

Also fixes a latent bug found on the way: TokenReal::clone() never copied
_datatype, so a cloned `1.0f` came back a double. Harmless while every real was
a double; not harmless with ddLDOUBLE live.

tests/testlongdouble.mad pins sizeof, a variable and a literal as varargs, two
long doubles in one varargs call, folded-vs-runtime agreement, full mantissa
retention from source text, and that long double is genuinely wider than
double. Every line matches gcc -std=c11 and clang -std=c11 exactly.

Not a fork change: MIR and c2mir already had everything needed.

+++ include/madc.h
@@ -2054,7 +2054,7 @@ protected:
     TokenBase *make_ident(const std::string &spelling);        // TokenIdent
     TokenBase *make_int(int64_t value);                        // TokenInt
     TokenBase *make_int(int64_t value, const std::string &src);// TokenInt + text
-    TokenBase *make_real(double value);                        // TokenReal
+    TokenBase *make_real(long double value);                   // TokenReal
     TokenBase *make_str(const std::string &bytes, bool wide = false); // TokenStr
     TokenBase *make_char(int code);                            // TokenChar
     TokenBase *make_datatype(const char *name, DataDef &dd);   // TokenDataType
